refactor: protected pageNumber form overflow

+++ Pulse.API/Features/Monitors/GetMonitors/GetMonitorsRequest.cs
@@ -2,5 +2,5 @@ using Pulse.BL.Features.Monitors;
 
 namespace Pulse.API.Features.Monitors.GetMonitors;
 
-public sealed record GetMonitorsRequest(MonitorStatus? Status, int? PageNumber, int? PageSize);
+public sealed record GetMonitorsRequest(MonitorStatus? Status, long? PageNumber, int? PageSize);
